requirements.txt
requests
discord.py
pyowm
time
colorama